services/ingestion-indexer/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pyyaml
numpy
pypdf
python-docx
